Change oldest upgradeable version to 3.0

+++ commentics/backend/model/login/login.php
@@ -38,10 +38,10 @@ class LoginLoginModel extends Model
     {
         $ip_address = $this->user->getIpAddress();
 
-        if ($this->db->numRows($this->db->query("SELECT * FROM `" . CMTX_DB_PREFIX . "attempts` WHERE `ip_address` = '" . $this->db->escape($ip_address) . "'"))) {
-            $this->db->query("UPDATE `" . CMTX_DB_PREFIX . "attempts` SET `amount` = `amount` + 1, `date_added` = NOW() WHERE `ip_address` = '" . $this->db->escape($ip_address) . "'");
+        if ($this->db->numRows($this->db->query("SELECT * FROM `" . CMTX_DB_PREFIX . "attempts` WHERE `type` = 'admin' AND `ip_address` = '" . $this->db->escape($ip_address) . "'"))) {
+            $this->db->query("UPDATE `" . CMTX_DB_PREFIX . "attempts` SET `amount` = `amount` + 1, `date_added` = NOW() WHERE `type` = 'admin' AND `ip_address` = '" . $this->db->escape($ip_address) . "'");
         } else {
-            $this->db->query("INSERT INTO `" . CMTX_DB_PREFIX . "attempts` SET `ip_address` = '" . $this->db->escape($ip_address) . "', `amount` = '1', `date_added` = NOW()");
+            $this->db->query("INSERT INTO `" . CMTX_DB_PREFIX . "attempts` SET `type` = 'admin', `ip_address` = '" . $this->db->escape($ip_address) . "', `amount` = '1', `date_added` = NOW()");
         }
 
         if ($this->db->numRows($this->db->query("SELECT * FROM `" . CMTX_DB_PREFIX . "admins` WHERE `username` = '" . $this->db->escape($username) . "'"))) {
@@ -53,8 +53,8 @@ class LoginLoginModel extends Model
     {
         $ip_address = $this->user->getIpAddress();
 
-        if ($this->db->numRows($this->db->query("SELECT * FROM `" . CMTX_DB_PREFIX . "attempts` WHERE `ip_address` = '" . $this->db->escape($ip_address) . "' AND `amount` >= 3"))) {
-            $query = $this->db->query("SELECT * FROM `" . CMTX_DB_PREFIX . "attempts` WHERE `ip_address` = '" . $this->db->escape($ip_address) . "' AND `amount` >= 3");
+        if ($this->db->numRows($this->db->query("SELECT * FROM `" . CMTX_DB_PREFIX . "attempts` WHERE `type` = 'admin' AND `ip_address` = '" . $this->db->escape($ip_address) . "' AND `amount` >= 3"))) {
+            $query = $this->db->query("SELECT * FROM `" . CMTX_DB_PREFIX . "attempts` WHERE `type` = 'admin' AND `ip_address` = '" . $this->db->escape($ip_address) . "' AND `amount` >= 3");
 
             $result = $this->db->row($query);
 
@@ -79,7 +79,7 @@ class LoginLoginModel extends Model
         $this->db->query("UPDATE `" . CMTX_DB_PREFIX . "admins` SET `last_login` = NOW() WHERE `id` = '" . (int) $admin_id . "'");
         $this->db->query("UPDATE `" . CMTX_DB_PREFIX . "admins` SET `resets` = '0' WHERE `id` = '" . (int) $admin_id . "'");
         $this->db->query("UPDATE `" . CMTX_DB_PREFIX . "admins` SET `login_attempts` = '0' WHERE `id` = '" . (int) $admin_id . "'");
-        $this->db->query("DELETE FROM `" . CMTX_DB_PREFIX . "attempts` WHERE `ip_address` = '" . $this->db->escape($ip_address) . "'");
+        $this->db->query("DELETE FROM `" . CMTX_DB_PREFIX . "attempts` WHERE `type` = 'admin' AND `ip_address` = '" . $this->db->escape($ip_address) . "'");
         $this->db->query("UPDATE `" . CMTX_DB_PREFIX . "logins` SET `date_modified` = NOW() ORDER BY `date_modified` ASC LIMIT 1");
     }
